move h2oFrameOutput to its own folder
