+++ src/routines/extendFrame.js
@@ -4,7 +4,7 @@ import { inspectTwoDimTable_ } from './inspectTwoDimTable_';
 import { inspectFrameColumns } from './inspectFrameColumns';
 import { inspectFrameData } from './inspectFrameData';
 
-import { h2oFrameOutput } from '../h2oFrameOutput';
+import { h2oFrameOutput } from '../h2oFrameOutput/h2oFrameOutput';
 
 import { flowPreludeFunction } from '../flowPreludeFunction';
 const flowPrelude = flowPreludeFunction();
